version 7

+++ src/app/customers/customers.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
 import {CustomerService} from "../services/customer.service";
-import {catchError, Observable, throwError} from "rxjs";
+import {catchError, map, Observable, throwError} from "rxjs";
 import {Customer} from "../model/customer.model";
 import {error} from "@angular/compiler-cli/src/transformers/util";
 import {FormBuilder, FormGroup} from "@angular/forms";
@@ -37,9 +37,18 @@ this.customers=this.customerService.searchCustomers(kw).pipe(
   }
 
   handleDeleteCustomer(c: Customer) {
+    let conf = confirm("Are you sure ?")
+    if (!conf) return;
     this.customerService.deleteCustomer(c.id).subscribe({
       next : (resp) => {
-        this.handleSearchCustomers();
+      this
+        this.customers=this.customers.pipe(
+          map(data=>{
+           let index= data.indexOf(c);
+           data.slice(index,1);
+            return data;
+          })
+        )
       },error :err => {
         console.log(err);
       }
